+++ lib/screens/bottom_nav/home/view/widget/discount_widget.dart
